+++ open-role.service/src/de/golesny/openrole/service/MailInfo.java
@@ -1,0 +1,8 @@
+package de.golesny.openrole.service;
+
+public class MailInfo {
+  public String email;
+  public String nick;
+  public String pw;
+  public String code;
+}