+++ open-role.service/src/de/golesny/openrole/service/util/DigestUtils.java
@@ -1,8 +1,10 @@
 package de.golesny.openrole.service.util;
 
+import java.math.BigInteger;
 import java.nio.charset.Charset;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
+import java.security.SecureRandom;
 
 import de.golesny.openrole.service.OpenRoleException;
 import de.golesny.openrole.service.OpenroleServiceServlet;
@@ -10,6 +12,7 @@ import de.golesny.openrole.service.OpenroleServiceServlet;
 public final class DigestUtils {
 
 	public static final Charset DEFAULTCHARSET = Charset.forName("UTF-8");
+	private static final String PWCHARS[] = {"abcdefghiklm","npqrstuvwxyz","ABCDEFGHIKLMN","PQRSTUVWXYZ","0123456789","!§%&?=/"};
 
 	public static String getSHA1(String plain) {
 		try {
@@ -26,5 +29,22 @@ public final class DigestUtils {
 			result += Integer.toString( ( b[i] & 0xff ) + 0x100, 16).substring( 1 );
 		}
 		return result;
+	}
+	
+	public static String createRandomString() {
+		SecureRandom random = new SecureRandom();
+		String newCode = new BigInteger(320, random).toString(32);
+		return newCode;
+	}
+	
+	public static String createPassword() {
+		SecureRandom random = new SecureRandom();
+		StringBuilder sb = new StringBuilder();
+		for (int i=0; i<8; i++) {
+			int slot = random.nextInt(PWCHARS.length);
+			char c = PWCHARS[slot].charAt(random.nextInt(PWCHARS[slot].length()));
+			sb.append(c);
+		}
+		return sb.toString();
 	}
 }