+++ webapp/Gruntfile.js
@@ -215,15 +215,15 @@ module.exports = function (grunt) {
     // By default, your `index.html`'s <!-- Usemin block --> will take care of
     // minification. These next options are pre-configured if you do not wish
     // to use the Usemin blocks.
-    // cssmin: {
-    //   dist: {
-    //     files: {
-    //       '<%= yeoman.dist %>/styles/main.css': [
-    //         '.tmp/styles/{,*/}*.css'
-    //       ]
-    //     }
-    //   }
-    // },
+    cssmin: {
+       dist: {
+         files: {
+           '<%= yeoman.dist %>/styles/main.css': [
+             '.tmp/styles/{,*/}*.css'
+           ]
+         }
+       }
+     },
     //uglify: {
     //   dist: {}	   
     // },
@@ -316,7 +316,7 @@ module.exports = function (grunt) {
           src: ['generated/*']
         }, {
           expand: true,
-          cwd: 'bower_components/bootstrap/dist',
+          cwd: 'app/bower_components/bootstrap/dist',
           src: 'fonts/*',
           dest: '<%= yeoman.dist %>'
         }]
@@ -357,7 +357,7 @@ module.exports = function (grunt) {
         noColor: false // If true, protractor will not use colors in its output.
       },
       run: {
-        configFile: "protractor.conf.js"
+        configFile: 'protractor.conf.js'
       }
     }
     
@@ -403,7 +403,7 @@ module.exports = function (grunt) {
     'ngmin',
     'copy:dist',
     //'cdnify',
-    //'cssmin',
+    'cssmin',
     'uglify',
     'filerev',
     'usemin',