+++ webapp/CHANGELOG.md
@@ -1,6 +1,14 @@
+## 2014-09-01
+
+Features:
+
+* Openrole
+    * User can now reset the password
+
 ## 2014-08-05
 
 Features:
+
 * CustomConf
     * New generic module that enables the user to create a custom module configuration
     * Configuration can be share with other users